Move Configuration and NotConfigured to confidence.models
